assessment/requirements.txt
flask>=3.0
python-dotenv
pydantic-settings>=2.0
numpy>=1.24
scikit-learn>=1.3
tiktoken>=0.5
openai>=1.40
anthropic>=0.40
google-generativeai>=0.4
requests>=2.31
gunicorn>=21.0
